ENH: use SearchVisitor in existing backend.search methods
