+++ superscore/backends/core.py
@@ -1,22 +1,12 @@
 """
 Base superscore data storage backend interface
 """
-import re
-from collections.abc import Container, Generator
-from typing import NamedTuple, Union
+from collections.abc import Generator
+from typing import Union
 from uuid import UUID
 
 from superscore.model import Entry, Root
-from superscore.type_hints import AnyEpicsType
-
-SearchTermValue = Union[AnyEpicsType, Container[AnyEpicsType], tuple[AnyEpicsType, ...]]
-SearchTermType = tuple[str, str, SearchTermValue]
-
-
-class SearchTerm(NamedTuple):
-    attr: str
-    operator: str
-    value: SearchTermValue
+from superscore.search_term import SearchTermType
 
 
 class _Backend:
@@ -66,41 +56,6 @@ class _Backend:
         - like (fuzzy match, depends on type of value)
         """
         raise NotImplementedError
-
-    @staticmethod
-    def compare(op: str, data: AnyEpicsType, target: SearchTermValue) -> bool:
-        """
-        Return whether data and target satisfy the op comparator, typically during application
-        of a search filter. Possible values of op are detailed in _Backend.search
-
-        Parameters
-        ----------
-        op: str
-            one of the comparators that all backends must support, detailed in _Backend.search
-        data: AnyEpicsType | Tuple[AnyEpicsType]
-            data from an Entry that is being used to decide whether the Entry passes a filter
-        target: AnyEpicsType | Tuple[AnyEpicsType]
-            the filter value
-
-        Returns
-        -------
-        bool
-            whether data and target satisfy the op condition
-        """
-        if op == "eq":
-            return data == target
-        elif op == "lt":
-            return data <= target
-        elif op == "gt":
-            return data >= target
-        elif op == "in":
-            return data in target
-        elif op == "like":
-            if isinstance(data, UUID):
-                data = str(data)
-            return re.search(target, data)
-        else:
-            raise ValueError(f"SearchTerm does not support operator \"{op}\"")
 
     @property
     def root(self) -> Root:
