+++ superscore/search_term.py
@@ -1,0 +1,12 @@
+from typing import Container, NamedTuple, Union
+
+from superscore.type_hints import AnyEpicsType
+
+SearchTermValue = Union[AnyEpicsType, Container[AnyEpicsType], tuple[AnyEpicsType, ...]]
+SearchTermType = tuple[str, str, SearchTermValue]
+
+
+class SearchTerm(NamedTuple):
+    attr: str
+    operator: str
+    value: SearchTermValue
