+++ superscore/tests/test_compare.py
@@ -4,12 +4,12 @@ from uuid import UUID
 
 import pytest
 
-from superscore.backends.core import SearchTerm
 from superscore.client import Client
 from superscore.compare import (AttributePath, DiffItem, EntryDiff,
                                 walk_find_diff)
 from superscore.model import (Collection, Entry, Parameter, Readback, Setpoint,
                               Severity, Snapshot, Status)
+from superscore.search_term import SearchTerm
 
 
 def simplify_path(path: AttributePath) -> AttributePath:
